Make default max leeway a integer

+++ src/Validator/Property/AbstractLeewayValidator.php
@@ -19,7 +19,7 @@ namespace RM\Standard\Jwt\Validator\Property;
 abstract class AbstractLeewayValidator implements PropertyValidatorInterface
 {
     final public const int DEFAULT_LEEWAY = 0;
-    final public const float DEFAULT_MAX_LEEWAY = 2 * 60;
+    final public const int DEFAULT_MAX_LEEWAY = 2 * 60;
     final public const int MINIMAL_LEEWAY = 0;
 
     /**
